+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [5.1.0-RC-7]
 ### Added
 - Error message on incompatible OkHttp dependency version
 - It's now possible to obtain Report Portal client by `Launch.currentLaunch().getClient()`